Refactored to remove result class
